+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * Converted non-static field initializers moved to constructor - [#281](https://github.com/icsharpcode/CodeConverter/issues/281)
 * Convert assignments using "Mid" built-in function
 * Improve conversion of array initializer types
+* Improve detection of compile-time constant case labels
 
 ### C# -> VB